safari-600.7-branch/LayoutTests/fast/dom/Window/Location/location-override-toString.html

<html>
<head>
    <script src="../../../../resources/js-test-pre.js"></script>
    <script src="resources/location-tests-functions.js"></script>
</head>
<body>
<script>
    window.location.toString = function() { return "haxored"; }

    var result = normalizeURL(String(window.location));
    var correctValue = normalizeURL(document.URL);
    shouldBe("result", "correctValue");
</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
</html>
